[refactor] Refactor function `add_validator_proposals` from runtime to epoch_manager
