+++ chain/epoch-manager/src/types.rs
@@ -2,18 +2,57 @@ use std::collections::{BTreeMap, HashMap};
 
 use borsh::{BorshDeserialize, BorshSerialize};
 
+use near_primitives::block_header::BlockHeader;
+use near_primitives::challenge::SlashedValidator;
 use near_primitives::epoch_manager::block_info::BlockInfo;
 use near_primitives::epoch_manager::epoch_info::EpochInfo;
 use near_primitives::hash::CryptoHash;
 use near_primitives::types::validator_stake::ValidatorStake;
 use near_primitives::types::{
-    AccountId, BlockHeight, EpochId, ShardId, ValidatorId, ValidatorStats,
+    AccountId, Balance, BlockHeight, EpochId, ShardId, ValidatorId, ValidatorStats,
 };
 use near_primitives::version::ProtocolVersion;
 
 use crate::EpochManager;
 
 pub type RngSeed = [u8; 32];
+
+/// Compressed information about block.
+/// Useful for epoch manager.
+#[derive(Default, Clone, Debug)]
+pub struct BlockHeaderInfo {
+    pub hash: CryptoHash,
+    pub prev_hash: CryptoHash,
+    pub height: BlockHeight,
+    pub random_value: CryptoHash,
+    pub last_finalized_height: BlockHeight,
+    pub last_finalized_block_hash: CryptoHash,
+    pub proposals: Vec<ValidatorStake>,
+    pub slashed_validators: Vec<SlashedValidator>,
+    pub chunk_mask: Vec<bool>,
+    pub total_supply: Balance,
+    pub latest_protocol_version: ProtocolVersion,
+    pub timestamp_nanosec: u64,
+}
+
+impl BlockHeaderInfo {
+    pub fn new(header: &BlockHeader, last_finalized_height: u64) -> Self {
+        Self {
+            hash: *header.hash(),
+            prev_hash: *header.prev_hash(),
+            height: header.height(),
+            random_value: *header.random_value(),
+            last_finalized_height,
+            last_finalized_block_hash: *header.last_final_block(),
+            proposals: header.validator_proposals().collect(),
+            slashed_validators: vec![],
+            chunk_mask: header.chunk_mask().to_vec(),
+            total_supply: header.total_supply(),
+            latest_protocol_version: header.latest_protocol_version(),
+            timestamp_nanosec: header.raw_timestamp(),
+        }
+    }
+}
 
 /// Aggregator of information needed for validator computation at the end of the epoch.
 #[derive(Clone, BorshSerialize, BorshDeserialize, Debug, Default)]
