+++ tools/state-viewer/src/commands.rs
@@ -9,10 +9,11 @@ use crate::{apply_chunk, epoch_info};
 use ansi_term::Color::Red;
 use near_chain::chain::collect_receipts_from_response;
 use near_chain::migrations::check_if_block_is_first_with_chunk_of_version;
+use near_chain::types::ApplyTransactionResult;
 use near_chain::types::RuntimeAdapter;
-use near_chain::types::{ApplyTransactionResult, BlockHeaderInfo};
 use near_chain::{ChainStore, ChainStoreAccess, ChainStoreUpdate, Error};
 use near_chain_configs::GenesisChangeConfig;
+use near_epoch_manager::types::BlockHeaderInfo;
 use near_epoch_manager::EpochManagerHandle;
 use near_epoch_manager::{EpochManager, EpochManagerAdapter};
 use near_primitives::account::id::AccountId;
@@ -642,7 +643,6 @@ pub(crate) fn print_chain(
 pub(crate) fn replay_chain(
     start_height: BlockHeight,
     end_height: BlockHeight,
-    home_dir: &Path,
     near_config: NearConfig,
     store: Store,
 ) {
@@ -654,12 +654,11 @@ pub(crate) fn replay_chain(
     let new_store = create_test_store();
     let epoch_manager =
         EpochManager::new_arc_handle(new_store.clone(), &near_config.genesis.config);
-    let runtime = NightshadeRuntime::from_config(home_dir, new_store, &near_config, epoch_manager);
     for height in start_height..=end_height {
         if let Ok(block_hash) = chain_store.get_block_hash_by_height(height) {
             let header = chain_store.get_block_header(&block_hash).unwrap().clone();
             println!("Height: {}, header: {:#?}", height, header);
-            runtime
+            epoch_manager
                 .add_validator_proposals(BlockHeaderInfo::new(
                     &header,
                     chain_store.get_block_height(header.last_final_block()).unwrap(),
